sensors: Add Bosch BMA4xx-series driver
This is a driver targetting the Bosch BMA 4-series accelerometers. It
has been specifically developed for the BMA422 but should be compatible
with others in that line, excepting the BMA400. Supports key attributes
and async RTIO one-shot operation. A sample application is included for
the nRF52840-DK board. I2C operation is supported, with stubs for a SPI
implementation provided for future improvement.

Signed-off-by: Tristan Honscheid <honscheid@google.com>